requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.0.1
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
packaging==25.0
passlib==1.7.4
pefile==2023.2.7
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
python-dotenv==1.1.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
